docs(api-reference): Stage 4 batch 3 — home Data Display + relocate config types
Home Label/Badge/ProgressBar/Gauge/ListView/DataTable/Tree/TreeNode into the
top-level `bootstack` API-reference page (new Data Display section, alphabetical)
and convert the 7 widget guides from a bottom `autoclass` to a table-only API
section that cross-links the reference.

Relocate `ColumnSpec`/`EditorType`/`FormOptions` definitions into
`widgets/types.py` (the low-level home, to avoid a circular import) and re-export
them from `bootstack.types`; `datatable.py`/`form.py` import from there. Drop
`EditorType` from the top-level `bootstack.__all__` — these are annotation/config
types, not compose-surface names. `ExportJob` stays local in `datatable.py`
(a return-handle Protocol, prose-only in the guide).

`on_*` payloads were already typed in the typing sweep (a6d6d496); audit confirmed
no changes needed. Updates `tests/test_public_surface.py`; clean docs build is
warning-free.
