+++ docs/widgets/listview.rst
@@ -235,9 +235,13 @@ Widget sizing
 API
 ---
 
-.. autoclass:: bootstack.widgets.listview.ListView
-   :members:
-   :undoc-members:
+The complete reference for :class:`ListView <bootstack.ListView>` lives on the
+:doc:`bootstack </api-reference/bootstack>` API page. At a glance:
+
+.. autosummary::
+   :nosignatures:
+
+   ~bootstack.ListView
 
 Full Example
 ------------
