+++ docs/widgets/progressbar.rst
@@ -129,9 +129,13 @@ Widget sizing
 API
 ---
 
-.. autoclass:: bootstack.widgets.progressbar.ProgressBar
-   :members:
-   :undoc-members:
+The complete reference for :class:`ProgressBar <bootstack.ProgressBar>` lives on
+the :doc:`bootstack </api-reference/bootstack>` API page. At a glance:
+
+.. autosummary::
+   :nosignatures:
+
+   ~bootstack.ProgressBar
 
 Full Example
 ------------
