+++ src/bootstack/__init__.py
@@ -101,7 +101,7 @@ from bootstack.widgets.tooltip import Tooltip
 
 from bootstack.widgets.calendar import Calendar
 from bootstack.widgets.form import Form
-from bootstack.widgets.form import FormItem, FieldItem, GroupItem, TabsItem, TabItem, EditorType
+from bootstack.widgets.form import FormItem, FieldItem, GroupItem, TabsItem, TabItem
 
 from bootstack.widgets.sidenav import (
     SideNav, SideNavItem, SideNavGroup, SideNavHeader, SideNavSeparator,
@@ -145,5 +145,4 @@ __all__ = [
     "Tooltip", "Toast", "toast",
     # Forms
     "Form", "FormItem", "FieldItem", "GroupItem", "TabsItem", "TabItem",
-    "EditorType",
 ]
